secure deleting

+++ levr_classes.py
@@ -615,7 +615,7 @@ def dealCreate(params,origin,upload_flag=True):
 		uid = params['uid']
 		
 		
-		deal = CustomerDeal(
+		deal = Deal(
 						parent			= uid,
 						is_exclusive	= False
 						)
@@ -640,7 +640,7 @@ def dealCreate(params,origin,upload_flag=True):
 	elif origin == 'admin_review':
 		#deal has already been uploaded by ninja - rewriting info that has been reviewed
 		dealID = enc.decrypt_key(params['deal'])
-		deal = CustomerDeal.get(db.Key(dealID))
+		deal = Deal.get(db.Key(dealID))
 		deal.been_reviewed		= True
 		deal.date_start			= datetime.now()
 		days_active				= int(params['days_active'])
@@ -753,7 +753,7 @@ class Customer(db.Model):
 	karma			= db.IntegerProperty(default=0)
 	new_notifications = db.IntegerProperty(default=0)
 	
-	#notifications, favorites, upvotes, etc...
+	# db references
 	followers		= db.ListProperty(db.Key) # Customer
 	favorites		= db.ListProperty(db.Key,default=[]) # Deal
 	upvotes			= db.ListProperty(db.Key,default=[]) # Deal
@@ -785,11 +785,11 @@ class Customer(db.Model):
 	email_friends		= db.ListProperty(str)
 	
 	#date stuff
-	date_created	= db.DateTimeProperty(auto_now_add=True)
-	date_last_edited= db.DateTimeProperty(auto_now=True)
-	date_last_login = db.DateTimeProperty(auto_now_add=True)
-	date_last_notified = db.DateTimeProperty(auto_now_add=True)
-	last_notified	= db.IntegerProperty(default=0)
+	date_created		= db.DateTimeProperty(auto_now_add=True)
+	date_last_edited	= db.DateTimeProperty(auto_now=True)
+	date_last_login 	= db.DateTimeProperty(auto_now_add=True)
+	date_last_notified 	= db.DateTimeProperty(auto_now_add=True)
+	last_notified		= db.IntegerProperty(default=0)
 	
 	#only for businesses - this property enables them to create offers for that business, without having verified it
 	#this property is only used up until they have linked a business
@@ -805,47 +805,100 @@ class Customer(db.Model):
 	new_redeem_count= db.IntegerProperty(default = 0) #number of unseen redemptions
 	vicinity		= db.StringProperty() #the area of the user, probably a college campus
 	
+	def merge_into_user(self,new_user):
+		'''
+		Merges two accounts. new_user will cannibalize the user that this method is called on.
+		This includes transferring all of the old users information.
+		
+		@param new_user:
+		@type new_user: Customer
+		'''
+		
+		new_user = self.secure_delete(self,new_user)
+		return new_user
 	
-	
-	def secure_delete(self):
+	def secure_delete(self,new_user=None):
 		'''
 		Securely deletes a Customer entity by removing all references to the entity in the database
 		and then calling db.Models delete method on the entity. See ya.
+		
+		If new_user is passed, then instead of deleting all relational references, they are transfered to the new user
+		
+		@param new_user: 
+		@type new_user: Customer
 		'''
-		key = self.key()
+		old_key = self.key()
+		new_key = new_user.key()
 		
 		#=======================================================================
 		# Notifications
 		# Delete the notifications
 		#=======================================================================
 		# to_be_notified notifications
-		notes = Notification.all().filter('to_be_notified',key).fetch(None)
+		to_be_notified = Notification.all().filter('to_be_notified',old_key).fetch(None)
 		# actor notifications
-		more_notes = self.notification_set.fetch(None)
-		notes.extend(more_notes)
+		actor = self.notification_set.fetch(None)
 		
-		logging.debug('Notifications deleted: {}'.format(notes.__len__()))
-		db.delete(notes)
+		notes = set([])
+		notes.update(to_be_notified)
+		notes.update(actor)
+		
+		
+		logging.debug('Notifications: {}'.format(to_be_notified.__len__()+actor.__len__()))
+		
+		
+		if new_user:
+			#transfer notifications
+			#exchange to_be_notified
+			for note in notes:
+				#user exists in to_be_notified
+				if old_key in user.to_be_notified:
+					#remove old reference
+					note.to_be_notified.remove(old_key)
+					#add new reference
+					note.to_be_notified.append(new_key)
+				# otherwise, user is the actor
+				else:
+					note.actor = new_user
+			db.put(notes)
+		else:
+			#delete all notifications
+			notes = set([])
+			notes.update(to_be_notified)
+			notes.update(actor)
+			db.delete(notes)
 		
 		#=======================================================================
 		# Other Customers
 		# Remove reference links
 		#=======================================================================
-		users = Customer.all().filter('followers',key).fetch(None)
+		users = Customer.all().filter('followers',old_key).fetch(None)
 		# remove the deleted users key
 		for user in users:
-			user.followers.remove(key)
+			#remove old reference
+			user.followers.remove(old_key)
+			#add new reference if there is one
+			if new_user:	user.followers.append(new_key)
+			
 		logging.debug(users)
-		logging.debug('Follower references removed: {}'.format(users.__len__()))
+		logging.debug('Follower references: {}'.format(users.__len__()))
 		db.put(users)
 		
 		#=======================================================================
 		# Deal
-		# have to delete the deals I guess... no other options since they are children
+		# 
 		#=======================================================================
-		deals = Deal.all().ancestor(key).fetch(None)
-		logging.debug('Child Deals removed: {}'.format(deals.__len__()))
-		db.delete(deals)
+		deals = Deal.all().ancestor(old_key).fetch(None)
+		logging.debug('Child Deals: {}'.format(deals.__len__()))
+		
+		for deal in deals:
+			if new_user:
+				#transfer ownership
+				deal.transfer_ownership_to(new_user)
+			else:
+				#secure deletion
+				deal.secure_delete()
+		
 		
 		
 		#=======================================================================
@@ -854,8 +907,10 @@ class Customer(db.Model):
 		#=======================================================================
 		businesses = self.businesses.fetch(None)
 		for business in businesses:
-			business.owner = None
-		logging.debug('Businesses no longer owned: {}'.format(businesses.__len__()))
+			#update or remove ownership
+			if new_user:	business.owner = new_user
+			else:			business.owner = None
+		logging.debug('Businesses ownership: {}'.format(businesses.__len__()))
 		db.put(businesses)
 		
 		
@@ -865,8 +920,10 @@ class Customer(db.Model):
 		#=======================================================================
 		reported_deals = self.reported_deals.fetch(None)
 		for deal in reported_deals:
-			deal.uid = None
-		logging.debug('ReportedDeals not longer being referenced: {}'.format(reported_deals.__len__()))
+			# Transfer or remove ownership
+			if new_user:	deal.uid = new_user
+			else:			deal.uid = None
+		logging.debug('ReportedDeals: {}'.format(reported_deals.__len__()))
 		db.put(reported_deals)
 		
 		#=======================================================================
@@ -875,10 +932,18 @@ class Customer(db.Model):
 		#=======================================================================
 		floating_content = self.floating_content.fetch(None)
 		logging.debug('Floating content removed: {}'.format(floating_content.__len__()))
-		db.delete(floating_content)
+		
+		if new_user:
+			# Update reference
+			for content in floating_content:
+				content.user = new_user
+			db.put(floating_content)
+		else:
+			# Delete the content... no longer relevant
+			db.delete(floating_content)
 		
 		#call the users delete function to delete the user
-		self.delete()
+#		self.delete()
 		
 		return
 	
@@ -1024,14 +1089,119 @@ class Deal(polymodel.PolyModel):
 	is_exclusive	= db.BooleanProperty(default=False)
 	
 	
+	def transfer_ownership_to(self,new_owner):
+		'''
+		Removes a deal
+		Transfers all of a deals information to a new deal owned by the new_owner
+		
+		@param new_owner: The new owner of the deal; entity, not the key
+		@type new_owner: Customer
+		'''
+		# get deal info
+		# set new owner
+		new_deal = Deal(parent=new_owner)
+		
+		logging.debug(log_dict(self.properties()))
+		
+		for key in self.properties():
+			# do not attempt to transfer private properties
+			if key[0] != '_':
+				val = getattr(self,key)
+				logging.debug('\n\n key: '+str(key)+'\n val: '+str(val))
+				setattr(new_deal, key, val)
+		
+		new_deal.put()
+		
+		self.secure_delete(new_deal)
+		return new_deal
 	
+	def secure_delete(self,new_deal=None):
+		'''
+		Securely deletes a deal entity, removing all of its references to other database entries
+		'''
+		old_key = self.key()
+		new_key = new_deal.key()
+		#=======================================================================
+		# Customer: upvotes, downvotes, favorites
+		#=======================================================================
+		upvotes = Customer.all().filter('upvotes',old_key).fetch(None)
+		downvotes = Customer.all().filter('downvotes',old_key).fetch(None)
+		favorites = Customer.all().filter('favorites',old_key).fetch(None)
+		
+		users = set([])
+		users.update(upvotes)
+		users.update(downvotes)
+		users.update(favorites)
+		
+		for user in users:
+			# upvotes
+			if old_key in upvotes:
+				user.upvotes.remove(old_key)
+				if new_deal:
+					user.upvotes.append(new_key)
+			# downvotes
+			if old_key in downvotes:
+				user.downvotes.remove(old_key)
+				if new_deal:
+					user.downvotes.append(new_key)
+			# favorites
+			if old_key in favorites:
+				user.favorites.remove(old_key)
+				if new_deal:
+					user.favorites.append(new_key)
+		logging.debug(users)
+		#replace users
+		db.put(users)
+		
+		#=======================================================================
+		# Notification
+		# If new deal is passed, replace the reference, otherwise delete the deal
+		#=======================================================================
+		notes = self.notifications.fetch(None)
+		
+		if new_deal:
+			for note in notes:
+				note.deal = new_deal
+			db.put(notes)
+		else:
+			db.delete(notes)
+		
+		#=======================================================================
+		# ReportedDeal
+		# If new deal is passed, replace the reference, otherwise delete
+		#=======================================================================
+		reported_deals = self.reported_deals.fetch(None)
+		
+		if new_deal:
+			for deal in reported_deals:
+				deal.deal = new_deal
+			db.put(reported_deals)
+		else:
+			db.delete(reported_deals)
+		
+		
+		#=======================================================================
+		# FloatingContent
+		# If new deal is passed, replace reference, otherwise delete
+		#=======================================================================
+		floating_content = self.floating_content.fetch(None)
+		
+		if new_deal:
+			for deal in floating_content:
+				deal.deal = new_deal
+			db.put(floating_content)
+		else:
+			db.delete(floating_content)
+		
+		self.delete()
+		
 	
 
 
 class CustomerDeal(Deal):
 #Sub-class of deal
 #A deal that has been uploaded by a user
-
+ 
 	gate_requirement= db.IntegerProperty(default = 5) #threshold of redeems that must be passed to earn a gate
 	gate_payment_per= db.IntegerProperty(default = 1) #dollar amount per gate
 	gate_count		= db.IntegerProperty(default = 0) #number of gates passed so far
@@ -1067,29 +1237,33 @@ class CustomerDeal(Deal):
 	
 
 class Notification(db.Model):
-	date			= db.DateTimeProperty(auto_now_add=True)
-	date_in_seconds	= db.IntegerProperty()
-	notification_type = db.StringProperty(required=True,choices=set(['upvote','followedUpload','newFollower','levelup','shared','levr','expired']))
-	line2			= db.StringProperty(default='')
-	to_be_notified	= db.ListProperty(db.Key)
-	deal			= db.ReferenceProperty(Deal,collection_name='notifications')
-	actor			= db.ReferenceProperty(Customer)
+	# Only has outbound references, no inbound
+	date				= db.DateTimeProperty(auto_now_add=True)
+	date_in_seconds		= db.IntegerProperty()
+	notification_type	= db.StringProperty(required=True,choices=set(['upvote','followedUpload','newFollower','levelup','shared','levr','expired']))
+	line2				= db.StringProperty(default='')
+	to_be_notified		= db.ListProperty(db.Key)
+	deal				= db.ReferenceProperty(Deal,collection_name='notifications')
+	actor				= db.ReferenceProperty(Customer)
 	
 
-class CashOutRequest(db.Model):
-#child of ninja
-	amount			= db.FloatProperty()
-	date_paid		= db.DateTimeProperty()
-	status			= db.StringProperty(choices=set(['pending','paid','rejected']))
-	payKey			= db.StringProperty()
-	money_available_paytime	= db.FloatProperty()
-	note			= db.StringProperty()
-	date_created	= db.DateTimeProperty(auto_now_add=True)
-	date_last_edited= db.DateTimeProperty(auto_now=True)
+#===============================================================================
+# class CashOutRequest(db.Model):
+# #child of ninja
+#	amount			= db.FloatProperty()
+#	date_paid		= db.DateTimeProperty()
+#	status			= db.StringProperty(choices=set(['pending','paid','rejected']))
+#	payKey			= db.StringProperty()
+#	money_available_paytime	= db.FloatProperty()
+#	note			= db.StringProperty()
+#	date_created	= db.DateTimeProperty(auto_now_add=True)
+#	date_last_edited= db.DateTimeProperty(auto_now=True)
+#===============================================================================
 
 class ReportedDeal(db.Model):
+	# Only has outbound references, no inbound
 	uid				= db.ReferenceProperty(Customer,collection_name='reported_deals')
-	dealID			= db.ReferenceProperty(Deal,collection_name='reported_deals')
+	deal			= db.ReferenceProperty(Deal,collection_name='reported_deals')
 	date_created	= db.DateTimeProperty(auto_now_add=True)
 	date_last_edited= db.DateTimeProperty(auto_now=True)
 	
@@ -1101,6 +1275,7 @@ class BusinessBetaRequest(db.Model):
 	date_created	= db.DateTimeProperty(auto_now_add=True)
 	
 class FloatingContent(db.Model):
+	#only has outbound references, no inbound
 	action				= db.StringProperty(required=True,choices=set(['upload','deal']))
 	contentID			= db.StringProperty(required=True)
 	user				= db.ReferenceProperty(Customer,collection_name='floating_content')
